feat(ui): add shared UpgradeToProDialog component
Refactored and centralized Pro upgrade dialog into a reusable shared
component. Replaced individual dialog instances across PlanManagement
and pricing plans with the new component for better code reuse and
consistency.
